% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generate_default_cost_matrix}
\alias{generate_default_cost_matrix}
\title{Generate a default cost matrix}
\usage{
generate_default_cost_matrix(data, delim = "")
}
\arguments{
\item{data}{DataFrame to be computed.}

\item{delim}{The delimiter separating atomic symbols.}
}
\value{
Cost matrix containing all possible characters in the raw data with all diagonal values set to 0 and others set to 1.
}
\description{
Generate a default cost matrix containing all possible characters in the raw data with all diagonal values set to 0 and others set to 1. This avoids constructing the matrix from scratch.
}
\examples{
df <- as.data.frame(rbind(a=c("ph_l_i_z","k_o_l"),b=c("b_l_i_s", "k_a:_l")))
default.cost <- generate_default_cost_matrix(df, "_")
}
