% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingdist-package.R
\docType{package}
\name{lingdist-package}
\alias{lingdist}
\alias{lingdist-package}
\title{lingdist: Fast Linguistic Distance and Alignment Computation}
\description{
A fast linguistic distance and alignment computation package. It implements generalized edit distance, Pointwise Mutual Information (PMI) distance, and Weighted Jaccard Distance (WJD). For generalized edit distance, the package allows users to define custom cost for every symbol's insertion, deletion, and substitution, and supports treating character combinations as single symbols (e.g., IPA with diacritics). It also provides detailed alignment information. PMI distance automatically learns costs from data. WJD is suitable for hierarchical categorical data with multiple forms. All functions are implemented in 'C++' and distance matrix computation is parallelized leveraging the 'RcppThread' package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fncokg/lingdist}
  \item Report bugs at \url{https://github.com/fncokg/lingdist/issues}
}

}
\author{
\strong{Maintainer}: Chao Kong \email{kongchao1998@gmail.com} (\href{https://orcid.org/0000-0002-6404-6142}{ORCID})

}
\keyword{internal}
