% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{long2squareform}
\alias{long2squareform}
\title{Convert long table to square form}
\usage{
long2squareform(data, symmetric = TRUE, default_diag = 0)
}
\arguments{
\item{data}{Dataframe in long table form. The first and second columns are labels and the third column stores the distance values.}

\item{symmetric}{Whether the distance matrix is symmetric (if cost matrix is not, then the distance matrix is also not).}

\item{default_diag}{The default value to fill in the diagonal positions. By default it is 0.0.}
}
\value{
Dataframe in square matrix form, rownames and colnames are labels. If the long table only contains \eqn{C_n^2} rows and `symmetric` is set to FALSE, then only lower triangle positions in the result are filled.
}
\description{
Convert a distance dataframe in long table form to a square matrix form. Values in the diagonal positions are automatically filled with `default_diag`. If you want self-defined diagonal values, define them in the long table.
}
\examples{
data <- as.data.frame(list(chars1=c("a","a","b"),chars2=c("b","c","c"),dist=c(1,2,3)))
mat <- long2squareform(data)
}
