% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pw_pmi_dist}
\alias{pw_pmi_dist}
\title{Compute PMI distance between all row pairs of a dataframe}
\usage{
pw_pmi_dist(
  data,
  delim = "",
  squareform = FALSE,
  parallel = FALSE,
  n_threads = 4L,
  max_epochs = 20L,
  tol = 1e-04,
  alignment_max_paths = 3L,
  quiet = FALSE
)
}
\arguments{
\item{data}{DataFrame with n rows and m columns indicating there are n languages or dialects to involve in the calculation and there are at most m words to base on, in which the rownames are the language ids.}

\item{delim}{The delimiter separating atomic symbols.}

\item{squareform}{Whether to return a dataframe in squareform.}

\item{parallel}{Whether to parallelize the computation.}

\item{n_threads}{The number of threads is used to parallelize the computation. Only meaningful if `parallel` is TRUE.}

\item{max_epochs}{Maximum number of epochs.}

\item{tol}{Tolerance for convergence.}

\item{alignment_max_paths}{Maximum number of paths to consider in alignment. There may be multiple optimal alignment paths between two strings; this parameter limits how many of them are considered when updating the cost matrix in each epoch.}

\item{quiet}{Whether to suppress all output messages.}
}
\value{
A list containing the following components:
\item{result}{A dataframe of distances, either in long table form or square form.}
\item{cost}{The final cost matrix used for distance calculation. Note: this is NOT the cost matrix after the last iteration, but the one before that, which is used to compute the final distances. That is, when you finished the iteration after 10 epochs, the cost matrix used to compute distances is actually the one being updated after the 9th epoch.}
\item{sum_diff}{The sum of absolute differences between the cost matrices of the last two iterations.}
\item{mean_diff}{The mean of absolute differences between the cost matrices of the last two iterations.}
}
\description{
Compute PMI (Pointwise Mutual Information) distance between all row pairs of a dataframe.
}
\examples{
df <- as.data.frame(rbind(a=c("ph_l_i_z","k_o_l"),b=c("b_l_i_s", "k_a:_l")))
result <- pw_pmi_dist(df, delim="_")
result <- pw_pmi_dist(df, delim="_", squareform=TRUE)
result <- pw_pmi_dist(df, delim="_", parallel=TRUE, n_threads=4)
}
